+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 
 ## [unreleased]
 ### Added
+  - Migration funcionality ([#LL-395](https://ht2labs.myjetbrains.com/youtrack/issue/LL-395))
+
 ### Changed
 ### Deprecated
 ### Removed